refactor: :recycle: repository route
