+++ apps/client/src/stores/log.ts
@@ -3,7 +3,7 @@ import { ref } from 'vue'
 import type { GetLogsQuery, Log } from '@cpn-console/shared'
 import { apiClient, extractData } from '@/api/xhr-client.js'
 
-export const useAdminLogStore = defineStore('admin-log', () => {
+export const useLogStore = defineStore('log', () => {
   const logs = ref<Log[]>([])
   const count = ref<number | undefined>(undefined)
