+++ apps/client/src/stores/project-repository.ts
@@ -9,14 +9,13 @@ export const useProjectRepositoryStore = defineStore('project-repository', () =>
   const repositories = ref<Repo[]>([])
 
   const getProjectRepositories = async (projectId: string) => {
-    repositories.value = await apiClient.Repositories.listRepositories({ params: { projectId } })
+    repositories.value = await apiClient.Repositories.listRepositories({ query: { projectId } })
       .then(response => extractData(response, 200))
   }
 
   const syncRepository = async (repositoryId: string, branchName: string) => {
-    if (!projectStore.selectedProject) throw new Error(projectMissing)
     await apiClient.Repositories.syncRepository({
-      params: { projectId: projectStore.selectedProject.id, repositoryId },
+      params: { repositoryId },
       body: { branchName },
     })
       .then(response => extractData(response, 204))
@@ -24,21 +23,21 @@ export const useProjectRepositoryStore = defineStore('project-repository', () =>
 
   const addRepoToProject = async (body: CreateRepositoryBody) => {
     if (!projectStore.selectedProject) throw new Error(projectMissing)
-    await apiClient.Repositories.createRepository({ body, params: { projectId: projectStore.selectedProject.id } })
+    await apiClient.Repositories.createRepository({ body, query: { projectId: projectStore.selectedProject.id } })
       .then(response => extractData(response, 201))
     await getProjectRepositories(projectStore.selectedProject.id)
   }
 
   const updateRepo = async (body: UpdateRepositoryBody & { id: RepositoryParams['repositoryId'] }) => {
     if (!projectStore.selectedProject) throw new Error(projectMissing)
-    await apiClient.Repositories.updateRepository({ body, params: { projectId: projectStore.selectedProject.id, repositoryId: body.id } })
+    await apiClient.Repositories.updateRepository({ body, params: { repositoryId: body.id } })
       .then(response => extractData(response, 200))
     await getProjectRepositories(projectStore.selectedProject.id)
   }
 
   const deleteRepo = async (repositoryId: string) => {
     if (!projectStore.selectedProject) throw new Error(projectMissing)
-    await apiClient.Repositories.deleteRepository({ params: { projectId: projectStore.selectedProject.id, repositoryId } })
+    await apiClient.Repositories.deleteRepository({ params: { repositoryId } })
       .then(response => extractData(response, 204))
     await getProjectRepositories(projectStore.selectedProject.id)
   }
