+++ apps/client/src/stores/zone.spec.ts
@@ -4,6 +4,9 @@ import { apiClient } from '../api/xhr-client.js'
 import { useZoneStore } from './zone.js'
 
 const apiClientGet = vi.spyOn(apiClient.Zones, 'listZones')
+const apiClientPost = vi.spyOn(apiClient.Zones, 'createZone')
+const apiClientPut = vi.spyOn(apiClient.Zones, 'updateZone')
+const apiClientDelete = vi.spyOn(apiClient.Zones, 'deleteZone')
 
 describe('Zone Store', () => {
   beforeEach(() => {
@@ -27,5 +30,38 @@ describe('Zone Store', () => {
 
     expect(apiClientGet).toHaveBeenCalledTimes(1)
     expect(zoneStore.zones).toMatchObject(zones)
+  })
+
+  it('Should create a new zone', async () => {
+    const adminZoneStore = useZoneStore()
+
+    const zone = { label: 'Zone à défendre', slug: 'zad' }
+    apiClientPost.mockReturnValueOnce(Promise.resolve({ status: 201, data: zone }))
+
+    await adminZoneStore.createZone(zone)
+
+    expect(apiClientPost).toHaveBeenCalledTimes(1)
+  })
+
+  it('Should update a zone', async () => {
+    const adminZoneStore = useZoneStore()
+
+    const zone = { id: 'zoneId', label: 'Zod à update' }
+    apiClientPut.mockReturnValueOnce(Promise.resolve({ status: 201, data: zone }))
+
+    await adminZoneStore.updateZone(zone.id, zone)
+
+    expect(apiClientPut).toHaveBeenCalledTimes(1)
+  })
+
+  it('Should delete a zone', async () => {
+    const adminZoneStore = useZoneStore()
+
+    const zoneId = 'zoneId'
+    apiClientDelete.mockReturnValueOnce(Promise.resolve({ status: 204, data: null }))
+
+    await adminZoneStore.deleteZone(zoneId)
+
+    expect(apiClientDelete).toHaveBeenCalledTimes(1)
   })
 })
