+++ apps/server/src/resources/repository/business.ts
@@ -1,6 +1,6 @@
-import type { Project, Repository, User } from '@prisma/client'
-import { type CreateRepositoryBody, type ProjectRoles, type UpdateRepositoryBody } from '@cpn-console/shared'
-import { addLogs, deleteRepository as deleteRepositoryQuery, getProjectInfos, getProjectInfosAndRepos, getUserById, initializeRepository, updateRepository as updateRepositoryQuery, getProjectRepositories as getProjectRepositoriesQuery } from '@/resources/queries-index.js'
+import type { Repository, User, Role } from '@prisma/client'
+import type { Project, CreateRepositoryBody, ProjectRoles, UpdateRepositoryBody } from '@cpn-console/shared'
+import { addLogs, deleteRepository as deleteRepositoryQuery, getRepositoryById as getRepositoryByIdQuery, getProjectInfosAndRepos, getUserById, initializeRepository, updateRepository as updateRepositoryQuery, getProjectRepositories as getProjectRepositoriesQuery, getProjectInfosOrThrow } from '@/resources/queries-index.js'
 import { checkInsufficientRoleInProject, checkRoleAndLocked } from '@/utils/controller.js'
 import { BadRequestError, DsoError, ForbiddenError, NotFoundError, UnauthorizedError, UnprocessableContentError } from '@/utils/errors.js'
 import { hook } from '@/utils/hook-wrapper.js'
@@ -38,14 +38,21 @@ export const getProjectAndCheckRole = async (
 }
 
 export const syncRepository = async (
-  projectId: Project['id'],
-  repositoryId: Repository['id'],
-  userId: User['id'],
-  branchName: string,
-  requestId: string,
-) => {
+  {
+    repositoryId,
+    userId,
+    branchName,
+    requestId,
+  }: {
+    repositoryId: Repository['id'],
+    userId: User['id'],
+    branchName: string,
+    requestId: string,
+  }) => {
   try {
-    await getProjectAndCheckRole(userId, projectId)
+    const repository = await getRepositoryByIdQuery(repositoryId)
+    const project = await getProjectInfosOrThrow(repository.projectId)
+    await checkUpsertRepository({ project, userId })
 
     const hookReply = await hook.misc.syncRepository(repositoryId, { branchName })
     await addLogs('Sync Repository', hookReply, userId, requestId)
@@ -59,28 +66,33 @@ export const syncRepository = async (
 }
 
 export const checkUpsertRepository = async (
-  userId: User['id'],
-  projectId: Project['id'],
-  minRole: ProjectRoles,
-) => {
-  const project = await getProjectInfos(projectId)
-  if (!project) throw new BadRequestError(`Le projet ayant pour id ${projectId} n'existe pas`)
+  {
+    project,
+    userId,
+    minRole = 'user',
+  }: { userId: User['id'], project: { locked: Project['locked'], roles: Role[] }, minRole?: ProjectRoles, }) => {
   const errorMessage = checkRoleAndLocked(project, userId, minRole)
-  if (errorMessage) throw new ForbiddenError(errorMessage, undefined)
+  if (errorMessage) throw new ForbiddenError(errorMessage)
 }
 
 export const createRepository = async (
-  projectId: Project['id'],
-  data: CreateRepositoryBody,
-  userId: User['id'],
-  requestId: string,
-) => {
-  await checkUpsertRepository(userId, projectId, 'owner')
+  {
+    projectId,
+    data,
+    userId,
+    requestId,
+  }: {
+    projectId: Project['id'],
+    data: CreateRepositoryBody,
+    userId: User['id'],
+    requestId: string
+  }) => {
   const user = await getUserById(userId)
   if (!user) throw new UnauthorizedError('Veuillez vous identifier')
 
   const project = await getProjectInfosAndRepos(projectId)
   if (!project) throw new BadRequestError(`Le projet ayant pour id ${projectId} n'existe pas`)
+  await checkUpsertRepository({ project, userId, minRole: 'owner' })
 
   if (project.repositories?.find(repo => repo.internalRepoName === data.internalRepoName)) throw new BadRequestError(`Le nom du dépôt interne ${data.internalRepoName} existe déjà en base pour ce projet`, undefined)
   const dbData = { ...data, projectId, isInfra: !!data.isInfra, isPrivate: !!data.isPrivate }
@@ -110,20 +122,26 @@ export const createRepository = async (
 }
 
 export const updateRepository = async (
-  projectId: Project['id'],
-  repositoryId: Repository['id'],
-  data: Partial<UpdateRepositoryBody>,
-  userId: User['id'],
-  requestId: string,
-) => {
-  const project = await getProjectInfos(projectId)
-  if (!project) throw new BadRequestError(`Le projet ayant pour id ${projectId} n'existe pas`)
-
-  const dbData = { ...data }
-  delete dbData.externalToken
-  const repo = await updateRepositoryQuery(repositoryId, dbData)
-
+  {
+    repositoryId,
+    data,
+    userId,
+    requestId,
+  }: {
+    repositoryId: Repository['id'],
+    data: Partial<UpdateRepositoryBody>,
+    userId: User['id'],
+    requestId: string,
+  }) => {
   try {
+    const repository = await getRepositoryByIdQuery(repositoryId)
+    const project = await getProjectInfosOrThrow(repository.projectId)
+    await checkUpsertRepository({ project, userId })
+
+    const dbData = { ...data }
+    delete dbData.externalToken
+    const repo = await updateRepositoryQuery(repositoryId, dbData)
+
     const { results } = await hook.project.upsert(project.id, {
       [repo.internalRepoName]: {
         username: repo.externalUserName ?? '',
@@ -142,16 +160,21 @@ export const updateRepository = async (
   }
 }
 
-export const deleteRepository = async (
-  projectId: Project['id'],
+export const deleteRepository = async ({
+  repositoryId,
+  userId,
+  requestId,
+}: {
   repositoryId: Repository['id'],
   userId: User['id'],
   requestId: string,
+},
 ) => {
-  const project = await getProjectAndCheckRole(userId, projectId, 'owner')
-
   try {
-    await checkUpsertRepository(userId, projectId, 'owner')
+    const repository = await getRepositoryByIdQuery(repositoryId)
+    const project = await getProjectInfosOrThrow(repository.projectId)
+
+    await checkUpsertRepository({ project, userId, minRole: 'owner' })
 
     await deleteRepositoryQuery(repositoryId)
