+++ apps/server/src/resources/user/controllers.spec.ts
@@ -1,7 +1,7 @@
 import prisma from '../../__mocks__/prisma.js'
 import { vi, describe, it, expect, beforeAll, afterEach, afterAll } from 'vitest'
-import { createRandomDbSetup, getRandomLog, getRandomPerm, getRandomRole, getRandomUser } from '@cpn-console/test-utils'
-import { adminGroupPath, projectIsLockedInfo } from '@cpn-console/shared'
+import { createRandomDbSetup, getRandomLog, getRandomPerm, getRandomRole, getRandomUser, repeatFn } from '@cpn-console/test-utils'
+import { adminGroupPath, projectIsLockedInfo, userContract } from '@cpn-console/shared'
 import { getConnection, closeConnections } from '../../connect.js'
 import { getRequestor, setRequestor } from '../../utils/mocks.js'
 import app from '../../app.js'
@@ -291,6 +291,87 @@ describe('User routes', () => {
       expect(response.statusCode).toEqual(400)
       expect(response.body).toBeDefined()
       expect(JSON.parse(response.body).error).toEqual('L\'utilisateur n\'est pas membre du projet')
+    })
+  })
+
+  // GET
+  describe('getUsersController', () => {
+    it('Should retrieve users', async () => {
+      // Create users
+      const users = repeatFn(5)(getRandomUser)
+      requestor.groups = [adminGroupPath]
+      setRequestor(requestor)
+
+      prisma.user.findMany.mockResolvedValue(users)
+      prisma.adminPlugin.findMany.mockResolvedValue([])
+
+      const response = await app.inject()
+        .get(userContract.getAllUsers.path)
+        .end()
+
+      expect(response.statusCode).toEqual(200)
+      expect(response.json()).toEqual(users.map(user => ({ ...user, isAdmin: false })))
+    })
+
+    it('Should return an error if retrieve users failed', async () => {
+      const errorMessage = 'Erreur de récupération des utilisateurs'
+      requestor.groups = [adminGroupPath]
+      setRequestor(requestor)
+
+      prisma.user.findMany.mockRejectedValue(new Error(errorMessage))
+
+      const response = await app.inject()
+        .get(userContract.getAllUsers.path)
+        .end()
+
+      expect(response.statusCode).toEqual(500)
+      expect(JSON.parse(response.body).error).toEqual(errorMessage)
+    })
+
+    it('Should return an error if requestor is not admin', async () => {
+      const requestor = getRandomUser()
+      requestor.groups = []
+      setRequestor(requestor)
+
+      const response = await app.inject()
+        .get(userContract.getAllUsers.path)
+        .end()
+
+      expect(response.statusCode).toEqual(403)
+      expect(JSON.parse(response.body).error).toEqual('Vous n\'avez pas les droits administrateur')
+    })
+  })
+
+  // PUT
+  describe('updateUserAdminRole', () => {
+    it('Should make an non-admin user admin', async () => {
+      const user = getRandomUser()
+      requestor.groups = [adminGroupPath]
+      setRequestor(requestor)
+
+      prisma.adminPlugin.findMany.mockResolvedValue([])
+
+      const response = await app.inject()
+        .put(`/api/v1/users/${user.id}`)
+        .body({ isAdmin: false })
+        .end()
+
+      expect(response.statusCode).toEqual(204)
+    })
+
+    it('Should make an admin user non-admin', async () => {
+      const user = getRandomUser()
+      requestor.groups = [adminGroupPath]
+      setRequestor(requestor)
+
+      prisma.adminPlugin.findMany.mockResolvedValue([])
+
+      const response = await app.inject()
+        .put(`/api/v1/users/${user.id}`)
+        .body({ isAdmin: true })
+        .end()
+
+      expect(response.statusCode).toEqual(204)
     })
   })
 })
